refactor(stats): lift masonry + DnD reorder out of StatsSection
Tier 3 architecture review (A2b), completing the StatsSection decomposition.

The card-layout engine (breakpoint columns, round-robin packing, span2
segments, the per-card DnD renderCard) moved to a StatsMasonry component;
the reorder interaction state (arranging / drag / hover + persist / moveId
/ reset) moved to a useCardReorder hook. StatsSection is now composition
only (372 → 148 lines including A2a). No behavior change.

Suite 474 green, lint clean, build OK.
